ai chat

+++ backend/app/models/user_comment_like.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import DateTime, func, ForeignKey, BigInteger
+from sqlalchemy import DateTime, func, ForeignKey, Integer
 from sqlalchemy.orm import Mapped, mapped_column
 from app.db.base import Base
 
@@ -7,12 +7,12 @@ from app.db.base import Base
 class UserCommentLike(Base):
     __tablename__ = "user_comment_likes"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True)
     comment_id: Mapped[int] = mapped_column(
-        BigInteger, ForeignKey("comments.id"), nullable=False
+        Integer, ForeignKey("comments.id"), nullable=False
     )
     user_id: Mapped[int] = mapped_column(
-        BigInteger, ForeignKey("users.id"), nullable=False
+        Integer, ForeignKey("users.id"), nullable=False
     )
     created_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), server_default=func.now(), nullable=False
